+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 
 # 1.0.1
 
--. Add arithmatic operation for Period, Instace, LocalDateTime, LocalDate, LocalTime 
+*   Add arithmatic operation for Period, Instace, LocalDateTime, LocalDate, LocalTime
+*   Remove DurationBuilder class
 
 # 1.0.0
 
 ### First release
 
--. Supply joda-time DateTime extension functions
+*   Supply joda-time DateTime extension functions